requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.1
autopep8==1.4.4
bootstrap4==0.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.6
django-crispy-forms==1.8.0
django-inline-actions==2.2.0
django-jet==1.0.8
django-mysql==3.2.0
django-registration-redux==2.6
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==20.0.0
h5py==2.10.0
idna==2.8
imutils==0.5.3
isort==4.3.21
joblib==0.14.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.4
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.2
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2